周课时采用小数

+++ tm-plan-api/grails-app/domain/cn/edu/bnuz/bell/planning/SchemeCourseDto.groovy
@@ -8,8 +8,8 @@ class SchemeCourseDto {
     String id
     String name
     BigDecimal credit
-    Integer theoryPeriod
-    Integer experimentPeriod
+    BigDecimal theoryPeriod
+    BigDecimal experimentPeriod
     Integer periodWeeks
     Boolean isPractical
     Integer educationLevel
